+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - change : add space around separator in tab filtered column
 - change : improve readme documentation
 - change : move github action (for github pages) to root
+- change : jsonjsdb use `__table__.json.js` instead of `__meta_.json.js`
 - fix : search bar transition z-index and left position
 - fix : page transition and search bar state management
 - fix : flash footer in header on page load with static rendering